seperated analysis from plots
